# backend/openapi.yaml
openapi: 3.0.3
info:
  title: May i Booking API
  version: 1.0.0
  description: >
    OpenAPI spec for May i Booking project.
    แบ่งหมวดตาม feature (Auth, Users, Books, Reviews, Cart, Addresses, Orders)
    และใช้ tag / x-roles เพื่อบอกว่า endpoint ไหนใช้โดย Guest / User / Admin
servers:
  - url: http://localhost:4000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
      required: [message]
    User:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Book:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer
        sku:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookCreateRequest:
      type: object
      properties:
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer
        sku:
          type: string
      required: [title, price, stock]
...
paths:
...
  /api/users/me:
    get:
      tags: [Users, User]
      summary: Get current user profile
      security:
        - BearerAuth: []
      x-roles: [user, admin]
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Users, User]
      summary: Update current user profile
      security:
        - BearerAuth: []
      x-roles: [user, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
  /api/users/me/favorites:
    get:
      tags: [Users, User]
      summary: Get favorite books of current user
      security:
        - BearerAuth: []
      x-roles: [user, admin]
      responses:
        '200':
          description: List of favorite book IDs
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorites:
                    type: array
                    items:
                      type: string
    put:
      tags: [Users, User]
      summary: Replace favorite books list
      security:
        - BearerAuth: []
      x-roles: [user, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                favorites:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Updated favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorites:
                    type: array
                    items:
                      type: string
  /api/users/me/avatar:
    post:
      tags: [Users, User]
      summary: Upload avatar for current user
      security:
        - BearerAuth: []
      x-roles: [user, admin]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Avatar uploaded successfully
    get:
      tags: [Users, User]
      summary: Get avatar image of current user
      security:
        - BearerAuth: []
      x-roles: [user, admin]
      responses:
        '200':
          description: Avatar image
          content:
            image/png: {}
            image/jpeg: {}
        '404':
          description: Avatar not found

  ################################
  #           BOOKS              #
  ################################
  /api/books:
    get:
      tags: [Books, Guest, User, Admin]
      summary: Get list of books
      x-roles: [guest, user, admin]
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'

    post:
      tags: [Books, Admin]
      summary: Create a new book (admin only)
      security:
        - BearerAuth: []
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Invalid input
        '403':
          description: Forbidden
...
  /api/books/{id}:
    get:
      tags: [Books, Guest, User, Admin]
      summary: Get book detail
      x-roles: [guest, user, admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found

    put:
      tags: [Books, Admin]
      summary: Update a book (admin only)
      security:
        - BearerAuth: []
      x-roles: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
      responses:
        '200':
          description: Book updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found

    delete:
      tags: [Books, Admin]
      summary: Delete a book (admin only)
      security:
        - BearerAuth: []
      x-roles: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Book deleted
        '404':
          description: Book not found
  /api/books/{id}/cover:
    get:
      tags: [Books, Guest, User, Admin]
      summary: Get book cover image
      x-roles: [guest, user, admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book cover image
          content:
            image/png: {}
            image/jpeg: {}
        '404':
          description: Cover not found

  ################################
  #          REVIEWS             #
  ################################
...